pages: move the page in the server side
